requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.1
pypdf==4.3.1
python-docx==1.1.2
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.5
httpx==0.27.2

